Refactor affected clusters table
Splits the table component into wrapper and table itself. The solution
helps to write cypress ct tests easier without intercepting network
requests.
