+++ src/AppEntry.js
@@ -11,18 +11,28 @@ const translations = {
   en: require('../compiled-lang/en.json'),
 };
 
-const AppEntry = ({ useLogger }) => (
+export const Intl = ({ children }) => (
   <IntlProvider
     locale={navigator.language.slice(0, 2)}
     defaultLocale="en"
     messages={translations[navigator.language.slice(0, 2)]}
     onError={console.error}
   >
+    {children}
+  </IntlProvider>
+);
+
+const AppEntry = ({ useLogger }) => (
+  <Intl>
     <Router basename={getBaseName(window.location.pathname, 3)}>
       <App useLogger={useLogger} />
     </Router>
-  </IntlProvider>
+  </Intl>
 );
+
+Intl.propTypes = {
+  children: PropTypes.element,
+};
 
 AppEntry.propTypes = {
   useLogger: PropTypes.bool,
